requirements.txt
click
python-dotenv
requests
